backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.9.0
httpx>=0.28.1
python-dotenv>=1.1.0
PyJWT==2.8.0
